requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.2.0
anthropic>=0.18.0
python-dotenv>=1.0.0
mcp>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
